backend/requirements.txt
websockets==12.0
asyncio
requests==2.31.0
dataclasses
web3==6.12.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
